Remove the bottom margin from block labels
When these are stacked, they need a bottom margin to set space between
them (10px, underneath).

When there is only one, or it’s the last one, the 10px bottom margin
should be removed.

This assumes there will be a parent element setting the spacing between
groups of form elements and the spacing will be set there.