Adjust styles for panels
Use box-sizing to prevent content set to 100% width overflowing the
panel for smaller viewports.

Set content to 100% width and float it left, so that it clears any
floated block labels before it.

Remove any bottom margin or padding, as any content inside the panel
should sit flush with the bottom of the panel.

Add a top margin, so that there is a gap between a form control and the
panel (but so that the two appear associated).